src/main/java/com/calculator/model/OperationModel.java
package com.calculator.model;

import java.math.BigDecimal;

public class OperationModel {
    private BigDecimal firstNumber;
    private BigDecimal secondNumber;

    public OperationModel() {
    }

    public OperationModel(BigDecimal firstNumber, BigDecimal secondNumber) {
        this.firstNumber = firstNumber;
        this.secondNumber = secondNumber;
    }

    public BigDecimal getFirstNumber() {
        return firstNumber;
    }

    public void setFirstNumber(BigDecimal firstNumber) {
        this.firstNumber = firstNumber;
    }

    public BigDecimal getSecondNumber() {
        return secondNumber;
    }

    public void setSecondNumber(BigDecimal secondNumber) {
        this.secondNumber = secondNumber;
    }
}